make rpmProblemGetLong uint64_t

+++ lib/rpmps.c
@@ -408,7 +408,7 @@ char * rpmProblemGetStr(rpmProblem prob)
     return(prob->str1);
 }
 
-unsigned long long rpmProblemGetLong(rpmProblem prob)
+uint64_t rpmProblemGetLong(rpmProblem prob)
 {
     return(prob->ulong1);
 }
